Enable Ginkgo solver on Windows
* allow to solve ginkgo with single precision
- use attribut fp32="true"
- no performance difference seen on macOS CPU
* change default preconditionder from amg to jacobi. Amg is very slow
  with 3D elasticity
* compile ginkgo with icx on win32 when using msvc
- introdude FIXED_C_CC to enforce compiler
- msvc's openMP is not sufficient for ginkgo
* patch ginkgo install, also required for non-OpenMP build
* enable ginkgo by default on windows
* fix a compiler warning
* avoid copying matrix data in ginkgo

+++ source/Optimization/Optimizer/PythonOptimizer.cc
@@ -582,7 +582,7 @@ PyObject* PythonOptimizer::Transfer(PyObject* args, bool derivative)
 
   int idx =  PyTuple_Size(args) == 2 ? PyLong_AsLong(PyTuple_GetItem(args,1)) : 0;
 
-  if(idx < 0 || idx > space->transfer.GetSize()-1)
+  if(idx < 0 || idx > (int) space->transfer.GetSize()-1)
     throw "invalid index for opt_(d_)_transfer() " + to_string(idx) + " range is 0 ... " + to_string(space->transfer.GetSize()-1);
 
   double val = PyFloat_AsDouble(PyTuple_GetItem(args,0));
